move consolidation logic to consolidation package and export
